# api/API.yaml
openapi: 3.0.3
info:
  version: 3.0.2
  title: Mop API
  description: "A list of the currently provided Interfaces to interact with the cluster, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local mop node main API
      port:
        default: "1633"
        description: Service port provided in mop node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local mop node main API
      port:
        default: "1633"
        description: Service port provided in mop node config
paths:
...
  "/feeds/{owner}/{topic}":
    post:
      summary: Create an initial feed root manifest
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "Common.yaml#/components/schemas/BNB Smart ChainAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "Common.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: type
          schema:
            $ref: "Common.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
        - $ref: "Common.yaml#/components/parameters/ClusterPinParameter"
        - $ref: "Common.yaml#/components/parameters/ClusterVoucherBatchId"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "Common.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "401":
          $ref: "Common.yaml#/components/responses/401"
        "402":
          $ref: "Common.yaml#/components/responses/402"
        "403":
          $ref: "Common.yaml#/components/responses/GatewayForbidden"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Find feed update
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "Common.yaml#/components/schemas/BNB Smart ChainAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "Common.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: at
          schema:
            type: integer
          required: false
          description: "Timestamp of the update (default: now)"
        - in: query
          name: type
          schema:
            $ref: "Common.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
      responses:
        "200":
          description: Latest feed update
          headers:
            "cluster-feed-index":
              $ref: "Common.yaml#/components/headers/ClusterFeedIndex"
            "cluster-feed-index-next":
              $ref: "Common.yaml#/components/headers/ClusterFeedIndexNext"
          content:
            application/json:
              schema:
                $ref: "Common.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "401":
          $ref: "Common.yaml#/components/responses/401"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SecurityToken